fix: user register bug fixes

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -13,6 +14,7 @@ class UserController extends Controller
     public function index(): Response
     {
         $users = User::query()
+            ->where('tenant_id', Auth::user()->tenant_id)
             ->when(request('search'), function ($query, $search) {
                 $query->where('name', 'like', "%{$search}%")
                     ->orWhere('email', 'like', "%{$search}%");
@@ -41,6 +43,10 @@ class UserController extends Controller
 
     public function edit(User $user): Response
     {
+        if ($user->tenant_id !== Auth::user()->tenant_id) {
+            abort(403);
+        }
+
         return Inertia::render('Users/Edit', [
             'user' => $user,
         ]);
@@ -48,6 +54,10 @@ class UserController extends Controller
 
     public function update(UserRequest $request, User $user): RedirectResponse
     {
+        if ($user->tenant_id !== Auth::user()->tenant_id) {
+            abort(403);
+        }
+
         $user->update($request->validated());
 
         return to_route('users.index')->with('success', 'Usuário atualizado com sucesso!');
@@ -55,6 +65,10 @@ class UserController extends Controller
 
     public function destroy(User $user): RedirectResponse
     {
+        if ($user->tenant_id !== Auth::user()->tenant_id) {
+            abort(403);
+        }
+
         $user->delete();
 
         return to_route('users.index')->with('success', 'Usuário excluído com sucesso!');
